Actualizar TiendaRepository.java

+++ idat/src/main/java/com/example/demo/Repository/TiendaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.Entity.TiendaMicas;
 
-public interface TiendaRepository extends JpaRepository<TiendaMicas, Long> {
+public interface TiendaRepository extends JpaRepository<TiendaMicas, Integer> {
     TiendaMicas findByNombre(String nombre);
     TiendaMicas findByNombreModelo(String nombreModelo);
     Optional<TiendaMicas> findById(int id);
